% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyness.R
\name{keyness_specificity}
\alias{keyness_specificity}
\title{Calclulates specificity}
\usage{
keyness_specificity(termCountsTarget, termCountsComparison, term)
}
\arguments{
\item{termCountsTarget}{A vector of counts}

\item{termCountsComparison}{A vector of counts}

\item{term}{A Term}
}
\value{
Specificiyt score
}
\description{
Calclulates specificity
}
