% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextProcessingBackend.R
\docType{data}
\name{QuantedaBackend}
\alias{QuantedaBackend}
\title{####################################################################################
#' Implementation of Quanteda Backend
#'
#' @author Janos Borst
#' @description Implementation Class for Quanteda Backend. Defines the functions that must be implemented
#' @format R6Class
#' @slot name Slot to set for the implemented backend.
#' @export}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
QuantedaBackend
}
\description{
####################################################################################
#' Implementation of Quanteda Backend
#'
#' @author Janos Borst
#' @description Implementation Class for Quanteda Backend. Defines the functions that must be implemented
#' @format R6Class
#' @slot name Slot to set for the implemented backend.
#' @export
}
\keyword{datasets}
