% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextProcessingBackend.R
\docType{data}
\name{SpacyrBackend}
\alias{SpacyrBackend}
\title{Implementation of Spacy Backend}
\format{R6Class}
\usage{
SpacyrBackend
}
\description{
Implementation Class for Tidytext Backend. Implements the pipline functions
}
\section{Slots}{

\describe{
\item{\code{name}}{Slot to set for the implemented backend.}
}}

\author{
Janos Borst
}
\keyword{datasets}
