% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextProcessingBackend.R
\docType{data}
\name{TextProcessingBackend}
\alias{TextProcessingBackend}
\title{Implementation of Textprocessing Backend. Abstraction Class.}
\format{R6Class}
\usage{
TextProcessingBackend
}
\description{
Abstraction Class for Backends. Defines the functions that must be implemented
Base class around different abtractions of textprocessing and analysis libraries like quanteda, or spacy
}
\section{Slots}{

\describe{
\item{\code{name}}{Slot to set for the implemented backend.}
}}

\author{
Janos Borst
}
\keyword{datasets}
